requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
langchain
langchain-core
langchain-community
langchain-groq
sentence-transformers
faiss-cpu
supabase